add test cases - correct logic for id 3125
