+++ src/test/resources/vardokFiles/2141.xml
@@ -1,0 +1,52 @@
+<FIMD xmlns="http://www.ssb.no/ns/fimd" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:dcterms="http//purl.org/dc/terms/" xmlns:xlink="http://www.w3.org/1999/xlink" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instanse" createdOn="2007-06-14" defaultValidFrom="2004-04-30" defaultValidTo="" id="urn:ssb:conceptvariable:vardok:2141" lastChangedDate="2010-04-27" otherLanguages="" type="ConceptVariable" xml:lang="nb" xsi:schemaLocation="http://www.ssb.no/ns/fimd Q:\DOK\FIMD\IT-analyse\Schemas\fimdv3.xsd">
+    <DC>
+        <dc:contributor>Seksjon for statistiske populasjoner</dc:contributor>
+        <dc:creator>Kari Helgeberg Olsen</dc:creator>
+        <dcterms:modified>2010-04-27</dcterms:modified>
+        <dcterms:valid>2004-04-30 - </dcterms:valid>
+        <dc:description>Denne variabelen finnes for hver etasje i bygget og sier hva slags etasje det er. Det finnes 4 mulige etasjetyper som har følgende definisjoner: Kjellerplan (kode K): Et plan der underkant dekke eller himling er høyst 0,75 m over planert terreng gjennomsnittsnivå rundt bygningen. Underetasje (kode U): Et plan der underkant dekke eller himling er høyere enn 0,75 m, men høyst 1,5 m over planert gjennomsnittsnivå. Hovedetasje (kode H): Et plan der underkant dekke er høyere enn 1,5 m over planert terrengs gjennomsnittsnivå rundt bygningen og der den frie bredden i høyde 1,9 m må være minst 1,9 m. Loftsetasje (kode L): Et tilgjengelig rom over øverste alminnelige etasje som ikke oppfyller kravene til alminnelig etasje. Fri høyde må være større enn eller lik 1,9 m i en bredde på minst 0,6 m. Loftsarealet måles til 0,6 m utenfor høyde på minst 1,9 m.</dc:description>
+        <dcterms:abstract> SSB-GAB/Matrikkel</dcterms:abstract>
+        <dcterms:tableOfContents>Matrikkel</dcterms:tableOfContents>
+        <dc:format>text/xml</dc:format>
+        <dc:identifier>urn:ssb:conceptvariable:vardok:2141</dc:identifier>
+        <dc:language>nb</dc:language>
+        <dc:publisher>Statistisk sentralbyrï¿½</dc:publisher>
+        <dc:rights>Statistiske metadata, produsert av Statistisk sentralbyrï¿½ (SSB) til produksjon av statistikk og annen alminnelig bruk.</dc:rights>
+        <dc:source/>
+        <dc:subject>10 - Næringsvirksomhet</dc:subject>
+        <dc:title>Etasjetype</dc:title>
+        <dc:type>Variabeldefinisjon</dc:type>
+    </DC>
+    <Common>
+        <Title>Etasjetype</Title>
+        <Description>Denne variabelen finnes for hver etasje i bygget og sier hva slags etasje det er. Det finnes 4 mulige etasjetyper som har følgende definisjoner: Kjellerplan (kode K): Et plan der underkant dekke eller himling er høyst 0,75 m over planert terreng gjennomsnittsnivå rundt bygningen. Underetasje (kode U): Et plan der underkant dekke eller himling er høyere enn 0,75 m, men høyst 1,5 m over planert gjennomsnittsnivå. Hovedetasje (kode H): Et plan der underkant dekke er høyere enn 1,5 m over planert terrengs gjennomsnittsnivå rundt bygningen og der den frie bredden i høyde 1,9 m må være minst 1,9 m. Loftsetasje (kode L): Et tilgjengelig rom over øverste alminnelige etasje som ikke oppfyller kravene til alminnelig etasje. Fri høyde må være større enn eller lik 1,9 m i en bredde på minst 0,6 m. Loftsarealet måles til 0,6 m utenfor høyde på minst 1,9 m.</Description>
+        <ContactPerson>
+            <CodeValue>kol</CodeValue>
+            <CodeText>Kari Helgeberg Olsen</CodeText>
+        </ContactPerson>
+        <ContactDivision>
+            <CodeValue>810</CodeValue>
+            <CodeText>Seksjon for statistiske populasjoner</CodeText>
+        </ContactDivision>
+        <Notes/>
+    </Common>
+    <Variable>
+        <InternalNotes/>
+        <StatisticalUnit>Matrikkel</StatisticalUnit>
+        <SubjectArea>
+            <CodeValue>10</CodeValue>
+            <CodeText>Næringsvirksomhet</CodeText>
+        </SubjectArea>
+        <ExternalSource/>
+        <InternalSource>SSB-GAB/Matrikkel</InternalSource>
+        <Sensitivity>Ordinï¿½r</Sensitivity>
+        <ExternalDocument/>
+        <DataElementName>ETASJE_TYPE</DataElementName>
+        <ShortNameWeb/>
+        <Calculation/>
+        <InternalDocument/>
+        <ExternalComment/>
+        <InternalReference/>
+    </Variable>
+    <Relations/>
+</FIMD>
